Move get_replication_slots method to Cluster
